% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{add_extrabytes}
\alias{add_extrabytes}
\alias{remove_attribute}
\title{Add attributes to a point cloud}
\usage{
add_extrabytes(data_type, name, description, scale = 1, offset = 0)

remove_attribute(name)
}
\arguments{
\item{data_type}{character. The data type of the extra bytes attribute. Can be "uchar", "char", "ushort",
"short", "uint", "int", "uint64", "int64", "float", "double".}

\item{name}{character. The name of the extra bytes attribute to add or remove to the file.}

\item{description}{character. A short description of the extra bytes attribute to add to the file (32 characters).}

\item{scale, offset}{numeric. The scale and offset of the data. See LAS specification. Leave unchanged
if not working with LAS files.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
Modify the memory layout of the point cloud to add an attributes to a point cloud. Values are
zeroed: the underlying point cloud is edited to support a new extrabyte attribute. This new
attribute can be populated later in another stage
}
\examples{
f <- system.file("extdata", "Example.las", package = "lasR")
fun <- function(data) { data$RAND <- runif(nrow(data), 0, 100); return(data) }
pipeline <- reader() +
  add_extrabytes("float", "RAND", "Random numbers") +
  callback(fun, expose = "xyz")
exec(pipeline, on = f)
}
