% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{add_rgb}
\alias{add_rgb}
\title{Add RGB attributes to a LAS file}
\usage{
add_rgb()
}
\description{
Modifies the LAS format to convert into a format with RGB attributes. The value are zeroed.
}
\examples{
f <- system.file("extdata", "example.las", package="lasR")

pipeline <- add_rgb() + write_las()
exec(pipeline, on = f)
}
