% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox2polys.R
\name{bbox2polys}
\alias{bbox2polys}
\title{create polygons from dataframe of bounding coordinates and ids}
\usage{
bbox2polys(idxxyy)
}
\arguments{
\item{idxxyy}{dataframe with 5 columns - id, minx,maxx, miny, maxy - representing the id, and bounding x,y of polygons}
}
\value{
SpatialPolygons object
}
\description{
create polygons from dataframe of bounding coordinates and ids
}
\details{
create polygons from dataframe of bounding coordinates and ids

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 March 08 Created\cr
}
}
\examples{
 #none yet

}
\seealso{
\code{\link{scan_las}}\cr \code{\link{read_las}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
