% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{chm}
\alias{chm}
\title{Canopy Height Model}
\usage{
chm(res = 1, tin = FALSE, ofile = tempfile(fileext = ".tif"))
}
\arguments{
\item{res}{numeric. The resolution of the raster.}

\item{tin}{bool. By default the CHM is a point-to-raster based methods i.e. each pixel is
assigned the elevation of the highest point. If \code{tin = TRUE} the CHM is a triangulation-based
model. The first returns are triangulated and interpolated.}

\item{ofile}{character. Full outputs are always stored on disk. If \code{ofile = ""}
then the stage will not store the result on disk and will return nothing. It will however
hold partial output results temporarily in memory. This is useful for stage that are only
intermediate stage.}
}
\description{
Create a Canopy Height Model using \link{triangulate} and \link{rasterize}.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
pipeline <- reader() + chm()
exec(pipeline, on = f)
}
\seealso{
\link{triangulate}
\link{rasterize}
}
