% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{classify_isolated_points}
\alias{classify_isolated_points}
\title{Classify noise points}
\usage{
classify_isolated_points(res = 5, n = 6L, class = 18L)
}
\arguments{
\item{res}{numeric. Resolution of the voxels.}

\item{n}{integer. The maximal number of 'other points' in the 27 voxels.}

\item{class}{integer. The class to assign to the points that match the condition.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
The stage identifies points that have only a few other points in their surrounding
3 x 3 x 3 = 27 voxels and edits the points to assign a target classification. Used with class 18,
it classifies points as noise. This stage modifies the point cloud in the pipeline but does not
produce any output.
}
