% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{classify_with_ptd}
\alias{classify_with_ptd}
\title{Ground Point Classification}
\usage{
classify_with_ptd(
  res = 10,
  angle = 30,
  distance = 2,
  filter = drop_noise(),
  spacing = 0.25,
  class = 2L
)
}
\arguments{
\item{res}{Scalar. The resolution for locating seed points. It takes the lowest point per grid cell.
It should be larger than any above ground structure such as building or bridges. In forested lands
10 m is good but larger values are recommended if human made structure can be observed. 50 m is
good in urban scene}

\item{angle}{Scalar. The threshold for iteration angle, as illustrated in Figure 3 of Axelsson's paper.
Use smaller values (close to 20) for flat areas and larger values (close to 40) for mountainous
areas.}

\item{distance}{Scalar. The threshold for iteration distance, as described in Figure 3 of Axelsson's paper.}

\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. For a given stage when a filter is applied, only the points that meet the criteria are processed.
The most common strings are \verb{Classification == 2"}, \code{"Z > 2"}, \code{"Intensity < 100"}. For more details see
\link{filters}.}

\item{spacing}{Scalar. The spacing of ground points. TIN densification halts if triangles are
smaller than this limit avoiding useless over-densification. The default setting is generally suitable
for most scenarios.}

\item{class}{Integer. The classification to assign to the points, typically 2 for ground points.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
Classify ground points using a modified version of the Progressive TIN
densification method by Axelsson (2000) (see references). The approach involves iteratively
constructing a triangulated surface model (TIN) to classify ground points (see details). \strong{This
algorithm is much more robust than the CSF in \link{classify_with_csf} and should be
preferred}. This methods also detect and thus classify low noise points i.e. points that are
part of ground point by construction but that are likely low outlier. Ground point are assigned class
2. Low outliers are assigned class 7.
}
\details{
The method begins by identifying local low points, assuming at least one ground-level point exists
within any X meters area. Initially, the triangles of the TIN model are mostly below the ground
surface, only touching lowest points at their vertices. As the iterations progress, additional
points are incorporated to more accurately follow the true ground surface.\cr\cr
The classification process is controlled by two main parameters:
\itemize{
\item Iteration angle: This controls the rate at which points are classified based on variations in
ground level. Use smaller values (e.g., 20) for flat terrain and larger values (e.g., 40) for
mountainous regions.
\item Iteration distance: This limits large upward jumps in the model, preventing the misclassification
of points in low vegetation or on small buildings.
}
The algorithm is robust to low noise points, meaning it can handle some level of noise without
needing explicit noise classification. However, it is preferable to have noise classified points
to guarantee optimal results.
\cr\cr
The densification process can also stop if the triangle edges become shorter than a specified length.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
pipeline = classify_with_ptd() + write_las()
ans = exec(pipeline, on = f, progress = TRUE)
}
\references{
Axelsson, P. (2000). DEM Generation from Laser Scanner Data Using adaptive TIN Models. International
Archives of Photogrammetry and Remote Sensing, 33(B4), 110–117.
https://www.isprs.org/proceedings/xxxiii/congress/part4/111_xxxiii-part4.pdf
}
