% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_plots.R
\name{clip_plots}
\alias{clip_plots}
\title{clip plots or polygons form lidar}
\usage{
clip_plots(idxyd = NA, idxy = NA, plot_polys = NA,
  id_field_plots = "plot", lasR_project = NA,
  lasR_project_polys = NA, plot_tile_intersect = NA, dir_out = NA,
  height = F, do_plot = F, return = F, n_core = 6, dir_dtm = NA,
  dir_las = NA, skip_existing = T, fix_dsm_bug = F)
}
\arguments{
\item{idxyd}{dataframe with point id, x and y coordinates, and circle diameter}

\item{idxy}{(optional) dataframe with polygon vertices (x,y) organized by point id}

\item{plot_polys}{(optional) shapefile path, or sp object with plot boundaries}

\item{id_field_plots}{(optional) in the event that a polygon is provided}

\item{lasR_project}{path to spatial polygons dataframe (intersection between dtm and las tiles)}

\item{lasR_project_polys}{spatial polygons dataframe (intersection between dtm and las tiles)}

\item{plot_tile_intersect}{path to spatial polygons dataframe representing intersection between plots and lasR_project}

\item{dir_out}{where to send clipped plots}

\item{height}{T/F subtract the ground}

\item{do_plot}{T/F make a plot of the intersections}

\item{return}{T/F return intersectiosn}

\item{n_core}{number of cores to use in clipping}

\item{dir_dtm}{in case path to dtms has changed from lasR_project}

\item{dir_las}{in case path to las has changed from lasR_project}

\item{skip_existing}{skip csv files that have been processed already - in case processing was interupted}
}
\value{
<Delete and Replace>
}
\description{
clip plots or polygons form lidar - requires that one first make a project using lasR_project which
 includes both lidar and dtms
}
\details{
There are three ways to provide the input shapes to this function - either as a
 list of plot ids, coordinates and plot diameters (4 column data.frame), as a list of vertices
 organized by plot_id (3 column data.frame), or as a polygon shapefile (path to shapefile) inwich case
 the plot id field must be specified "id_field_plots".

\cr
Revision History
\tabular{ll}{
1.0 \tab date and revisions.. \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{another_function}}\cr \code{\link{yet_another_function}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
