% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_plots_fusion.R
\name{clip_plots_fusion}
\alias{clip_plots_fusion}
\title{clip plots or polygons form lidar}
\usage{
clip_plots_fusion(
  idxyd = NA,
  dir_las = NA,
  dir_dtm = NA,
  dir_clipdata = "c:\\\\fusion\\\\clipdata.exe",
  dir_out = NA,
  out_f = c(".las", ".laz"),
  clipdata_switches = c("/height /shape:1", "")[1],
  n_core = 6,
  temp = "c:\\\\temp\\\\clipdata",
  run = T
)
}
\value{
<Delete and Replace>
}
\description{
clip plots or polygons form lidar - requires that one first make a project using lasR_project which
 includes both lidar and dtms
}
\details{
There are three ways to provide the input shapes to this function - either as a
 list of plot ids, coordinates and plot diameters (4 column data.frame), as a list of vertices
 organized by plot_id (3 column data.frame), or as a polygon shapefile (path to shapefile) inwich case
 the plot id field must be specified "id_field_plots".

\cr
Revision History
\tabular{ll}{
1.0 \tab date and revisions.. \cr
}
}
\examples{

cd = read.csv("D:\\\\data\\\\wadnr_hood_canal\\\\field_data\\\\hood_canal_locations.csv")
cd1 = data.frame(cd[!is.na(cd[,"x"]),c("plot_id","x","y")], width = 37.2*2 )


res=clip_plots_fusion(
  idxyd=cd1
  ,dir_las = "D:\\\\data\\\\wadnr_hood_canal\\\\las\\\\hood_canal_6in_DSM_2015\\\\"
  ,dir_dtm = "D:\\\\data\\\\lidar_wa_dtms\\\\"
  ,dir_clipdata="c:\\\\fusion\\\\clipdata.exe"
  ,dir_out = "D:\\\\data\\\\wadnr_hood_canal\\\\plot_clips\\\\6in"
  ,out_f = c(".las",".laz")[1]
  ,clipdata_switches="/height /shape:1"
  ,n_core=6
  ,temp = "c:\\\\temp\\\\clipdata"
  ,run=T
)


}
\seealso{
\code{\link{another_function}}\cr \code{\link{yet_another_function}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
