% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sqlite.R
\name{csv_to_sqlite}
\alias{csv_to_sqlite}
\title{load csvs into database}
\usage{
csv_to_sqlite(db, csv_folder, tb_summary = "gm_summary", tb_csv = "gm",
  project = "naip_2015", resolution = "66", units = "feet",
  proj4 = "", notes = "", skip_loaded = T, n_load = NA,
  use_col_classes = T)
}
\arguments{
\item{db}{sqlite database connection}

\item{csv_folder}{location of csv files}

\item{tb_summary}{output name of summary table of csv files}

\item{tb_csv}{table to dump csv files into}

\item{project}{name of project}

\item{resolution}{resolution of input data}

\item{units}{unites of resolution}

\item{proj4}{proj4 string of coordinate data}

\item{notes}{any relevant notes}

\item{skip_loaded}{should csv files which are already loaded be skipped}

\item{n_load}{max number of files to load}

\item{use_col_classes}{read colClasses from first file, and apply to remaining files - big read speedup}
}
\description{
point at a directory and load fusion derived tabular csv data into an sqlite database
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 1/18/2018 header added \cr
}
}
\examples{
 library(lasR)
 dir_csv="I:\\\\projects\\\\2017_WA_DSM_Pilot\\\\2017Aug_NAIP\\\\gridmetrics1\\\\gridmetrics_csv\\\\"
   dir_sqlite="I:\\\\projects\\\\2017_WA_DSM_Pilot\\\\2017Aug_NAIP\\\\sqlite\\\\"
   dir.create(dir_sqlite)

   require(RSQLite)
   db <- dbConnect(SQLite(), dbname=file.path(dir_sqlite,"2015_NAIP_Metrics2.db"),sychronous = "off")


   csv_to_sqlite(db=db
                 ,csv_folder=dir_csv
   )

}
\seealso{
\code{\link{run_gridmetrics}}\cr \code{\link{lasR_project}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
