% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{delete_points}
\alias{delete_points}
\title{Filter and delete points}
\usage{
delete_points(filter = "")
}
\arguments{
\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. The available filters are those from LASlib and can be found by running \link{filter_usage}.
For a given algorithm when a filter is applied, only the points that meet the criteria are processes.
The most common strings are "-keep_first", "-keep_class 2", "drop_z_below 2". For more details see
\link{filters}.}
}
\description{
Remove some points from the point cloud. This stage modifies the point cloud in the pipeline
but does not produce any output.
}
\examples{
f <- system.file("extdata", "Megaplot.las", package="lasR")
read <- reader(f)
filter <- delete_points(keep_z_above(4))

pipeline <- read + summarise() + filter + summarise()
processor(pipeline)
}
