% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{dsm}
\alias{dsm}
\alias{chm}
\title{Digital Surface Model}
\usage{
dsm(res = 1, tin = FALSE, ofile = tempfile(fileext = ".tif"))

chm(res = 1, tin = FALSE, ofile = tempfile(fileext = ".tif"))
}
\arguments{
\item{res}{numeric. The resolution of the raster.}

\item{tin}{bool. By default the DSM is a point-to-raster based methods i.e. each pixel is
assigned the elevation of the highest point. If \code{tin = TRUE} the CHM is a triangulation-based
model. The first returns are triangulated and interpolated.}

\item{ofile}{character. Full outputs are always stored on disk. If \code{ofile = ""}
then the stage will not store the result on disk and will return nothing. It will however
hold partial output results temporarily in memory. This is useful for stage that are only
intermediate stage.}
}
\description{
Digital Surface Model using \link{triangulate} and \link{rasterize}. \code{chm()} is a alias to \code{dsm()}
but is misleading because it actually computes a DSM and does not include a normalization step.
\code{chm()} is deprecated. Technically \code{dsm(tin = TRUE)} is simply the association of the stage \code{triangulate()}
and \code{rasterize()} while \code{dsm(tin = FALSE)} is simply an alias to \code{rasterize("max")}.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
pipeline <- reader() + dsm()
exec(pipeline, on = f)
}
\seealso{
\link{triangulate}
\link{rasterize}
}
