% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{dtm}
\alias{dtm}
\title{Digital Terrain Model}
\usage{
dtm(res = 1, add_class = NULL, ofile = tempfile(fileext = ".tif"))
}
\arguments{
\item{res}{numeric. The resolution of the raster.}

\item{add_class}{integer. By default it triangulates using ground points (class 2). It is possible
to provide additional classes such as 9 for water.}

\item{ofile}{character. Full outputs are always stored on disk. If \code{ofile = ""}
then the stage will not store the result on disk and will return nothing. It will however
hold partial output results temporarily in memory. This is useful for stage that are only
intermediate stage.}
}
\description{
Create a Digital Terrain Model using \link{triangulate} and \link{rasterize}.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
pipeline <- reader(f) + dtm()

}
\seealso{
\link{triangulate}
\link{rasterize}
}
