% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{edit_attribute}
\alias{edit_attribute}
\title{Edit an attribute of the points}
\usage{
edit_attribute(filter = "", attribute = "", value = 0)
}
\arguments{
\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. For a given stage when a filter is applied, only the points that meet the criteria are processed.
The most common strings are \verb{Classification == 2"}, \code{"Z > 2"}, \code{"Intensity < 100"}. For more details see
\link{filters}.}

\item{attribute}{string. The name of an attribute to edit}

\item{value}{numeric. The value to assign. Be careful, if the user try to assign a value out of
range of representable value for a given data type it will be clamped.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
Edit an attribute of the points by filtering the point based on criteria.
}
\examples{
f <- system.file("extdata", "Example.las", package="lasR")

edit = edit_attribute(filter = c("Z < 975", "Z > 974"), attribute = "UserData", value = 2)
io = write_las(templas())
pipeline = edit + io
ans = exec(pipeline, on = f)
}
