% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_metrics.R
\name{experimental_metrics}
\alias{experimental_metrics}
\title{compute experimental lidar metrics}
\usage{
experimental_metrics(x = NA, y = NA, z = NA, i = NA, rn = NA,
  cl = NA, r = NA, g = NA, b = NA, resxy = 20, resxyz = 20,
  ressurf = 5, htcover = 6, zthresh = c(3, 6, 12, 20),
  adjustz = c("none", "positive", "min0", "zq01"), outlier = c(min = NA, max
  = NA))
}
\arguments{
\item{x}{lidR object X vector : las1$X}

\item{y}{lidR object Y vector : las1$Y}

\item{z}{lidR object Z vector : las1$Z}

\item{i}{lidR object intensity vector : las1$Intensity}

\item{rn}{lidR object return vector : las1$ReturnNumber}

\item{cl}{lidR object return vector : las1$Classification}

\item{r}{lidR object red vector : las1$R}

\item{g}{lidR object green vector : las1$G}

\item{b}{lidR object blue vector : las1$B}

\item{resxy}{resolution to use when computing xy statistic - grid resolution}

\item{resxyz}{resolution to use when computing xyz statistic - voxel resolution}

\item{ressurf}{resolution to use when computing raster areas - grid resolution}

\item{htcover}{height to use when computing cover and threshold for eliminating ground returns}

\item{zthresh}{height thresholds to use for height ratios round(100*length(z[ z > zthresh[i]]) / length(z[ z > htcover]),2)}

\item{adjustz}{(optional) method used to shift z values up/down. When there is no DTM or the DTM is poor (e.g. when used with DAP) this can help}

\item{outlier}{a two value vector with minimum and maximum values used to restrict the range of z values considered}
}
\value{
list of statistics, currently only computed on x,y,z
}
\description{
take advantage of xy and 3d information
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 4/4/2019 \cr
}
}
\examples{
   if(!"las1" \%in\% ls()) las1 = readLAS("D:\\\\Box\\\\sync\\\\R\\\\analyses\\\\wa_dsm_env_fia\\\\data\\\\plot_clips_ht\\\\plot_44.laz")
   test2 = lasmetrics(las1,experimental_metrics(X,Y,Z))
   test2

}
\seealso{
\code{\link{another_function}}\cr \code{\link{yet_another_function}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
