% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{filter_with_grid}
\alias{filter_with_grid}
\title{Select highest or lowest points}
\usage{
filter_with_grid(res, operator = "min", filter = "")
}
\arguments{
\item{res}{numeric. The resolution of the grid}

\item{operator}{string. Can be min or max to retain lowest or highest points}

\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. For a given stage when a filter is applied, only the points that meet the criteria are processed.
The most common strings are \verb{Classification == 2"}, \code{"Z > 2"}, \code{"Intensity < 100"}. For more details see
\link{filters}.}
}
\description{
Select and retained only highest or lowest points per grid cell
}
