% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\alias{keep_class}
\alias{drop_class}
\alias{keep_first}
\alias{drop_first}
\alias{keep_ground}
\alias{keep_ground_and_water}
\alias{drop_ground}
\alias{keep_noise}
\alias{drop_noise}
\alias{keep_z_above}
\alias{drop_z_above}
\alias{keep_z_below}
\alias{drop_z_below}
\alias{drop_duplicates}
\alias{filter_usage}
\alias{print.laslibfilter}
\alias{+.laslibfilter}
\title{Point filters}
\usage{
keep_class(x)

drop_class(x)

keep_first()

drop_first()

keep_ground()

keep_ground_and_water()

drop_ground()

keep_noise()

drop_noise()

keep_z_above(x)

drop_z_above(x)

keep_z_below(x)

drop_z_below(x)

drop_duplicates()

filter_usage()

\method{print}{laslibfilter}(x, ...)

\method{+}{laslibfilter}(e1, e2)
}
\arguments{
\item{x}{numeric or integer as a function of the filter used.}

\item{...}{Unused.}

\item{e1, e2}{lasR objects.}
}
\description{
lasR uses LASlib/LASzip, the library developed by Martin Isenburg to read and write LAS/LAZ files.
Thus, the flags that are available in \code{LAStools} are also available in \code{lasR}.
Filters are strings to put in the \code{filter} arguments of the \code{lasR} algorithms. The list of available
strings is accessible with \code{filter_usage}. For convenience, the most useful filters have an associated
function that returns the corresponding string.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
filter_usage()
gnd = keep_class(c(2,9))
reader_las(gnd)
triangulate(filter = keep_ground())
rasterize(1, "max", filter = "-drop_z_below 5")
}
