% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{focal}
\alias{focal}
\title{Calculate focal ("moving window") values for each cell of a raster}
\usage{
focal(raster, size, fun = "mean", ofile = temptif())
}
\arguments{
\item{raster}{LASRalgorithm. A stage that produces a raster.}

\item{size}{numeric. The window size **in the units of the point cloud**, not in pixels. For example, 2 means 2 meters
or 2 feet, not 2 pixels.}

\item{fun}{string. Function to apply. Supported functions are 'mean', 'median', 'min', 'max', 'sum'.}

\item{ofile}{character. Full outputs are always stored on disk. If \code{ofile = ""}
then the stage will not store the result on disk and will return nothing. It will however
hold partial output results temporarily in memory. This is useful for stage that are only
intermediate stage.}
}
\value{
This stage produces a raster. The path provided to `ofile` is expected to be `.tif` or any
other format supported by GDAL.
}
\description{
Calculate focal ("moving window") values for each cell of a raster using various functions. NAs
are always omitted; thus, this stage effectively acts as an NA filler. The window is always circular.
The edges are handled by adjusting the window.
}
\examples{
f <- system.file("extdata", "Topography.las", package = "lasR")

chm = rasterize(2, "zmax")
chm2 = lasR:::focal(chm, 8, fun = "mean")
chm3 = lasR:::focal(chm, 8, fun = "max")
pipeline <- reader() + chm + chm2 + chm2
ans = exec(pipeline, on = f)

terra::plot(ans[[1]])
terra::plot(ans[[2]])
terra::plot(ans[[3]])
}
