% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{geometry_features}
\alias{geometry_features}
\title{Compute pointwise geometry features}
\usage{
geometry_features(k, r = NULL, features = "")
}
\arguments{
\item{k}{Integer. Number of nearest neighbors (k-nearest neighbors).}

\item{r}{Numeric. Radius limit to search for the k-nearest neighbors.}

\item{features}{String. Geometric feature to export. Each feature is added into an extrabyte
attribute. Use 'C' for the 9 principal component coefficients, 'E' for the 3 eigenvalues of the
covariance matrix, 'a' for anisotropy, 'p' for planarity, 's' for sphericity, 'l' for linearity,
'o' for omnivariance, 'c' for curvature, 'e' for the sum of eigenvalues, 'i' for the angle
(inclination in degrees relative to the azimuth), and 'n' for the 3 components of the normal vector.
Notice that the uppercase labeled components allow computing all the lowercase labeled components.
Default is "". In this case, the singular value decomposition is computed but serves no purpose.}
}
\description{
Compute pointwise geometry features based on k-nearest neighbors. Each feature is added into an
extrabyte attribute. The names of the extrabytes attributes (if recorded) are \code{coeff00}, \code{coeff01},
\code{coeff02} and so on, \code{lambda1}, \code{lambda2}, \code{lambda3}, \code{anisotropy}, \code{planarity}, \code{sphericity}, \code{linearity},
\code{omnivariance}, \code{curvature}, \code{eigensum}, \code{angle}, \code{normalX}, \code{normalY}, \code{normalZ}. There is a total
of 23 attributes that can be added. All the features are recorded with single precision floating points.
It is strongly discouraged to use them all. This stage modifies the point cloud in the pipeline
but does not produce any output.
}
\examples{
f <- system.file("extdata", "Example.las", package = "lasR")
pipeline <- geometry_features(8, features = "pi") + write_las()
ans <- exec(pipeline, on = f)
}
\references{
Hackel, T., Wegner, J. D., & Schindler, K. (2016). Contour detection in unstructured 3D
point clouds. In Proceedings of the IEEE conference on computer vision and pattern recognition (pp. 1610-1618).
}
