% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_header_double}
\alias{get_header_double}
\title{get_header_double}
\usage{
get_header_double(header, mnem, default_val)
}
\arguments{
\item{header}{A DataFrame containing header data}

\item{mnem}{A string containing the desired mnemonic}

\item{default_val}{A double to be used if the value can't be found or converted}
}
\value{
A double containing the value
}
\description{
A function to get a double value from a header
}
