% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_header_int}
\alias{get_header_int}
\title{get_header_int}
\usage{
get_header_int(header, mnem, default_val)
}
\arguments{
\item{header}{A DataFrame containing header data}

\item{mnem}{A string containing the desired mnemonic}

\item{default_val}{An int to be used if the value can't be found or converted}
}
\value{
An int containing the value
}
\description{
A function to get an int value from a header
}
