% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_header_string}
\alias{get_header_string}
\title{get_header_string}
\usage{
get_header_string(header, mnem)
}
\arguments{
\item{header}{A DataFrame containing header data}

\item{mnem}{A string containing the desired mnemonic}
}
\value{
A string containing the value
}
\description{
A function to get a string value from a header
}
