% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proj4.R
\name{get_proj4}
\alias{get_proj4}
\title{Search for proj4 string}
\usage{
get_proj4(details = c("HARN", "wash", "ftUS", "nad83"), EPSG,
  omit = "deprecated", ignore.case = T, fixed = T, ...)
}
\arguments{
\item{details}{strings to use in searching for projection nam}

\item{EPSG}{(optional) dataframe of EPSGs as from make_EPSG()}

\item{omit}{strings not desired - e.g. "ft" when looking for meters}

\item{...}{}
}
\value{
data frame of candidates
}
\description{
helper to parse results form EPSG() in RGDAL and get desired proj4 string
}
\details{
helper to parse results form EPSG() in RGDAL and get desired proj4 string

\cr

Revision History
\tabular{ll}{
1.0 \tab 2015 Jan 05 Roxygen header added \cr
}
}
\examples{

get_proj4(c("NAD83(HARN)","Washington","ft"))
get_proj4(c("NAD83(HARN)","Washington","South","ft"))


}
\seealso{
\code{\link{rgdal}}\cr \code{\link{EPSG}}\cr
}
\author{
Jacob Strunk
}
