% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLogData}
\alias{getLogData}
\title{Get Log Data}
\usage{
getLogData(lines, curveNames, nullVal)
}
\arguments{
\item{lines}{Vector of LAS text lines}

\item{curveNames}{Vector of curve names}

\item{nullVal}{The null value}
}
\value{
A data.frame of parsed header values
}
\description{
Gets log data from an unwrapped LAS file
}
