% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{normalize}
\alias{normalize}
\alias{hag}
\title{Height Above Ground}
\usage{
normalize()

hag()
}
\description{
Normalize the point cloud using \link{triangulate} and \link{transform_with}. It triangulates the
ground points and then applies \code{transform_with} to linearly interpolate the elevation of each point
within each triangle. \code{normalize()} modifies the Z elevation and record the Height Above Ground (HAG)
effectively flattening the topography and normalizing the point cloud. \code{hag()} in another hand record
the HAG in an extrabytes attributes names 'HAD' and preserves the original Height Above See Z coordinates
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
pipeline <- reader_las() + normalize() + write_las()
exec(pipeline, on = f)
}
\seealso{
\link{triangulate}
\link{transform_with}
}
