% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/laser.crr.R
\name{laser.crr}
\alias{laser.crr}
\title{Canopy Relief Ratio (CRR)}
\usage{
laser.crr(z, id)
}
\arguments{
\item{z}{laser echo heights}
}
\value{
data.frame with FR, RR, IR, BL according to the notation in the reference
}
\description{
Compute canopy relief ratio, a quantitative measure of the relative shape of the canopy.
}
\examples{
data(las2)
laser.crr (las2$z,las2$ID)
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}
\references{
e.g. Mura, M., McRoberts, R. E., Chirici, G., & Marchetti, M. (2015). Estimating and mapping forest structural diversity using airborne laser scanning data. Remote Sensing of Environment, 170, 133–142.
}

