% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laser.fractional.R
\name{laser.fractional}
\alias{laser.fractional}
\title{LiDAR models of fractional cover}
\arguments{
\item{z}{laser echo heights}

\item{i}{laser echo intensity}

\item{r}{laser return number}

\item{n}{laser number of returns}

\item{c}{laser classification 2 = ground}

\item{id}{id separate metrics will be computed for each id}

\item{gtv}{ground threshold value default 1.3}
}
\value{
data.frame with FR, RR, IR, BL according to the notation in the reference
}
\description{
Compute the four lidar-based models of canopy fractional cover proposed by Hopkinson and Chasmer (2009)
}
\examples{
data(las2)
laser.fractional(las2$z,las2$i,las2$r,las2$n,las2$c,las2$ID)
}
\references{
Hopkinson, C., & Chasmer, L. (2009). Testing LiDAR models of fractional cover across multiple forest ecozones. Remote Sensing of Environment, 113(1), 275–288.
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}
