% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laser.metrics.R
\name{laser.metrics}
\alias{laser.metrics}
\title{Compute the height and density variables from point cloud data}
\arguments{
\item{x}{surface height}

\item{id}{unique idenfitier for which variables will be aggregated}

\item{gtv}{ground threshold value. Default is 2}

\item{ctv}{canopy threshold value. Default is 95 percentile}

\item{percentiles}{the percentiles to be computed. Defaults to seq(0.1,0.9,0.1)}

\item{prefix}{optional prefix of varaible names}

\item{suffix}{optional suffix of varaible names}

\item{distpre}{indicator of distribution metrics. Default is "H"}

\item{denspre}{indicator of density metrics. Default is "D"}
}
\value{
data.frame with the numbers and proportion of echoes in differnt categories.
}
\description{
Compute the height and density variables from point cloud data according to Naesset 2004.
}
\note{
Date: Nov 2, 2016
}
\examples{
data(las)
#First returns
las1 <- subset(las,r==1)
Vars1 <- laser.metrics(las1$z,las1$ID,gtv=1.3,suffix='.F')
head(Vars1)

# Last returns uten single! 
las2 <- subset(las,r==n & r != 1) 
Vars2 <- laser.metrics(las2$z,las2$ID,gtv=1.3,suffix='.L')

head(Vars2)
}
\references{
Næsset, E., & Gobakken, T. (2008). Estimation of above- and below-ground biomass across regions of the boreal forest zone using airborne laser. Remote Sens. Environ., 112, 3079-3090
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}
