% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laser.returns.R
\name{laser.returns}
\alias{laser.returns}
\title{Compute the number and proportion of returntypes}
\usage{
laser.returns(dz, RetNum, NumRet, id)
}
\arguments{
\item{dz}{surface height}

\item{RetNum}{return numb}

\item{NumRet}{number of returns}

\item{id}{unique idenfitier for which variables will be aggregated}
}
\value{
data.frame with the numbers and proportion of echoes in differnt categories.
}
\description{
Read the header of a las-file into R
}
\examples{
data(las)
prop <- laser.returns(las$dz,las$r,las$n,id=las$ID)
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}
