% Generated by roxygen2 (4.0.2): do not edit by hand
\name{laser.returntype}
\alias{laser.returntype}
\title{Classify returntypes, Echo categories}
\usage{
laser.returntype(RetNum, NumRet)
}
\arguments{
\item{RetNum}{return numb}

\item{NumRet}{number of returns}
}
\value{
vector with echo categories "SINGLE" (first and only), "FIRST" (first of many), "LAST"(last of many), "INTER" (intemediate), "ZERO" (labeling errors).
}
\description{
Read the header of a las-file into R
}
\examples{
data(las)
Echo.cat <- laser.returntype(las$r,las$n)
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}

