% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasR.R
\name{lasR}
\alias{lasR}
\title{stuff}
\usage{
lasR()
}
\value{
#stuff
}
\description{
stuff
}
\details{
stuff

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 April 14 \cr
}
}
\examples{
#stuff

}
\seealso{
\code{\link{lidR}}\cr \code{\link{rLIDAR}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
