% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_lasR_project.R, R/lasR_project.R
\name{lasR_project}
\alias{lasR_project}
\title{Build Lidar / DEM tiling project}
\usage{
lasR_project(
  dir_las = NA,
  dir_dtm = NA,
  dir_project = "c:/lidar_projects/",
  project = "test_project",
  project_dtm = "test_project",
  project_las = "test_project",
  dtm_year = "2099",
  las_year = "2099",
  do_scan_dtms = T,
  do_scan_las = T,
  tile_size = 1650,
  pixel_size = 66,
  xmn = 561066,
  xmx = 2805066,
  ymn = 33066,
  ymx = 1551066,
 
    crs = "+proj=lcc +lat_1=47.33333333333334 +lat_2=45.83333333333334 +lat_0=45.33333333333334 +lon_0=-120.5 +x_0=500000.0001016001 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs",
  mask = NA,
  return = F
)

lasR_project(
  dir_las = NA,
  dir_dtm = NA,
  dir_project = "c:/lidar_projects/",
  project = "test_project",
  project_dtm = "test_project",
  project_las = "test_project",
  dtm_year = "2099",
  las_year = "2099",
  do_scan_dtms = T,
  do_scan_las = T,
  tile_size = 1650,
  pixel_size = 66,
  xmn = 561066,
  xmx = 2805066,
  ymn = 33066,
  ymx = 1551066,
 
    crs = "+proj=lcc +lat_1=47.33333333333334 +lat_2=45.83333333333334 +lat_0=45.33333333333334 +lon_0=-120.5 +x_0=500000.0001016001 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs",
  mask = NA,
  return = F
)
}
\arguments{
\item{dir_las}{where are las files}

\item{dir_dtm}{where are dtm files}

\item{dir_project}{where to place project}

\item{project}{project name}

\item{project_dtm}{dtm project name}

\item{project_las}{las project name}

\item{dtm_year}{year of dtm files}

\item{las_year}{year of las files}

\item{tile_size}{processing tile size}

\item{pixel_size}{raster pixel size}

\item{xmn, xmx, ymn, ymx}{bound box for processing grid}

\item{crs}{projection string}

\item{scan_dtms}{?scan dtm files}

\item{scan_las}{?scan las files}
}
\value{
<Delete and Replace>

<Delete and Replace>
}
\description{
Scans lidar and dems and finds out where they intersect

Scans lidar and dems and finds out where they intersect
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 3/28/2017 Created \cr
}

<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 3/28/2017 Created \cr
1.1 \tab 3/17/2020 force users to use either geopackage or shapefile \cr
}
}
\examples{

 lasR_project(
  dir_las="C:\\\\Temp\\\\las_test\\\\"
  ,dir_dtm="C:\\\\Temp\\\\dtm_test\\\\"
  ,dir_project="C:\\\\Temp\\\\naip_2015_t1650_p66\\\\"
  ,project="test_project"
  ,project_dtm="some_project"
  ,project_las="some_project"
  ,dtm_year="2099"
  ,dlas_year="2099"
  ,scan_dtms=T
  ,scan_las=T
  ,create_project=T
  ,tile_size=1650
  ,pixel_size=66
  ,xmn=561066,xmx=2805066,ymn=33066,ymx=1551066
  ,crs="+proj=lcc +lat_1=47.33333333333334 +lat_2=45.83333333333334 +lat_0=45.33333333333334 +lon_0=-120.5 +x_0=500000.0001016001 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs"
)



 lasR_project(
  dir_las="C:\\\\Temp\\\\las_test\\\\"
  ,dir_dtm="C:\\\\Temp\\\\dtm_test\\\\"
  ,dir_project="C:\\\\Temp\\\\naip_2015_t1650_p66\\\\"
  ,project="test_project"
  ,project_dtm="some_project"
  ,project_las="some_project"
  ,dtm_year="2099"
  ,dlas_year="2099"
  ,scan_dtms=T
  ,scan_las=T
  ,create_project=T
  ,tile_size=1650
  ,pixel_size=66
  ,xmn=561066,xmx=2805066,ymn=33066,ymx=1551066
  ,crs="+proj=lcc +lat_1=47.33333333333334 +lat_2=45.83333333333334 +lat_0=45.33333333333334 +lon_0=-120.5 +x_0=500000.0001016001 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs"
)


}
\seealso{
\code{\link{scan_dtm}}\cr \code{\link{scan_las}}\cr

\code{\link{scan_dtm}}\cr \code{\link{scan_las}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>

Jacob Strunk <Jstrunk@fs.fed.us>
}
