% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasR_subset.R
\name{lasR_subset}
\alias{lasR_subset}
\title{subset tiles in a lasR project}
\usage{
lasR_subset(path_project, path_shapefile, dir_out = "", ignore_proj4 = T,
  return = F, plot = T)
}
\arguments{
\item{path_project}{path to project file}

\item{path_shapefile}{path to shapefile}

\item{dir_out}{where to put subset}

\item{ignore_proj4}{should proj4 differences be ignored}

\item{return}{should the subset project be returned}

\item{plot}{show a plot of project subset}
}
\description{
supply a polygon and subset tiles in a lasR project
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 April 14 Implemented \cr
}
}
\examples{
 lasR_proj_small=lasR_subset("C:\\\\projects\\\\2017_WA_DSM_Pilot\\\\DSM_Pilot_A_lasR\\\\intersections.csv"
 ,"C:\\\\projects\\\\2017_WA_DSM_Pilot\\\\boundary\\\\extentB.shp"
 ,dir_out="C:\\\\projects\\\\2017_WA_DSM_Pilot\\\\DSM_Pilot_A_lasR\\\\subset_project\\\\"
 )

}
\seealso{
\code{\link{lasR_project}}\cr \code{\link{scan_las}}\cr \code{\link{scan_dtm}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
