% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.las.R
\name{list.to.dt}
\alias{list.to.dt}
\title{list.to.dt}
\usage{
list.to.dt(l)
}
\arguments{
\item{l}{A list}
}
\value{
Nothing, data.frames are converted in place
}
\description{
Converts all data.frames in a list to data.tables in place
}
\keyword{internal}
