% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{logdataParser}
\alias{logdataParser}
\title{Parse log data}
\usage{
logdataParser(log_dat, curves, nodata, wrap)
}
\arguments{
\item{log_dat}{character. Raw section containing well log curve data}

\item{curves}{dataframe. Curve information}

\item{nodata}{numeric. Nodata value read from the well information section}
}
\value{
dataframe. Well log data.
}
\description{
Parse log data
}
