% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmp.R
\name{ncores}
\alias{ncores}
\alias{half_cores}
\title{Determine the number of available cores}
\usage{
ncores()

half_cores()
}
\value{
An integer representing the number of available CPU cores.
}
\description{
This function returns the number of available CPU cores that can be utilized.
}
