% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{normalize}
\alias{normalize}
\title{Normalize the point cloud}
\usage{
normalize(extrabytes = FALSE)
}
\arguments{
\item{extrabytes}{bool. If FALSE the coordinate Z of the point cloud is modified and becomes the
height above ground (HAG). If TRUE the coordinate Z is not modified and a new extrabytes attribute
named 'HAG' is added to the point cloud.}
}
\description{
Normalize the point cloud using \link{triangulate} and \link{transform_with}
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
pipeline <- reader_las() + normalize() + write_las()
exec(pipeline, on = f)
}
\seealso{
\link{triangulate}
\link{transform_with}
}
