% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{parametersParser}
\alias{parametersParser}
\title{Parse the parameter section of an LAS ile}
\usage{
parametersParser(param_dat)
}
\arguments{
\item{param_dat}{character. Raw parameter section}
}
\value{
dataframe. Parameter information
}
\description{
Parse the parameter section of an LAS ile
}
