% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_curves}
\alias{parse_curves}
\title{parse_curves}
\usage{
parse_curves(
  lines,
  curve_names,
  format,
  delim = " \\t",
  null_str = "-999.25",
  first_line = 0L,
  last_line = -1L
)
}
\arguments{
\item{lines}{A vector of strings containing log curve data}

\item{curve_names}{A vector of strings containing the names of the log curves}

\item{format}{A vector of strings containing the formats of the log curves}

\item{delim}{A string specifying the delimiter}

\item{null_str}{A string indicating what value to replace with NA}

\item{first_line}{An integer indicating the index of the first line in lines containing log data}

\item{last_line}{An integer indicating the last index of lines containing log data.}
}
\value{
A dataframe containing the curves
}
\description{
A function to parse delimited curve data from a LAS file.
}
