% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_header}
\alias{parse_header}
\title{parse_header}
\usage{
parse_header(lines, first_line = 0L, last_line = -1L)
}
\arguments{
\item{lines}{A vector of strings containing header data.}

\item{first_line}{An integer indicating the index of the first line in lines containing header data.}

\item{last_line}{An integer indicating the last index of lines containing header data.}
}
\value{
A dataframe containing the header data.
}
\description{
A function to parse header lines from a LAS file.
}
