% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{plot.lasrcloud}
\alias{plot.lasrcloud}
\title{Plot Method for 'lasrcloud' Objects}
\usage{
\method{plot}{lasrcloud}(x, ...)
}
\arguments{
\item{x}{An object of class 'lasrcloud'.}

\item{...}{Additional arguments (not used).}
}
\description{
This function defines a custom plot method for objects of class 'lasrcloud'.
}
\examples{
\dontrun{
f <- system.file("extdata", "Topography.las", package="lasR")
las <- read_cloud(f)
las
plot(las)
}
}
