% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2polys.R
\name{points2polys}
\alias{points2polys}
\title{create polygons from dataframe of vertices and ids}
\usage{
points2polys(idxy)
}
\arguments{
\item{idxy}{dataframe with 3 columns - id, x, y - representing the id, x, and y of vertices of polygons}
}
\value{
SpatialPolygons object
}
\description{
create polygons from dataframe of vertices and ids
}
\details{
create polygons from dataframe of vertices and ids

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 March 08 Created\cr
}
}
\examples{
 none yet

}
\seealso{
\code{\link{xy2FIApoly}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
