% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{print.lasrcloud}
\alias{print.lasrcloud}
\title{Print Method for 'lasrcloud' Objects}
\usage{
\method{print}{lasrcloud}(x, ...)
}
\arguments{
\item{x}{An object of class 'lasrcloud'.}

\item{...}{Additional arguments (not used).}
}
\description{
This function defines a custom print method for objects of class 'lasrcloud'.
}
