% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr.R
\name{read_cloud}
\alias{read_cloud}
\title{Read a point cloud in memory}
\usage{
read_cloud(file)
}
\arguments{
\item{file}{a file containing a point cloud. Currently only LAS and LAZ files are supported}
}
\description{
Read a point cloud in memory. The point cloud is stored in a C++ data structure
and is not exposed to users
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
las <- read_cloud(f)
las
u = exec(chm(5), on = las)
u
}
