% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dtm.R
\name{read_dtm}
\alias{read_dtm}
\alias{read_dtm_header}
\title{read USDA FUSION format (Bob McGaughey) raster files and headers}
\usage{
read_dtm(dtm_files = NA, dir_dtm = NA, pattern = "[.]dtm$", plot = F,
  NA_val = -1, return_polygon = F)

read_dtm_header(dtm_file, dir_dtm = NA, pattern = "[.]dtm$",
  recursive = TRUE, out_file = NA)
}
\arguments{
\item{dir_dtm}{path to directory containing FUSION dtm files}

\item{pattern}{extension for FUSION dtm files}

\item{plot}{with 'dtm_file' optionally plot dtm with rgl and rasterVis}

\item{NA_val}{recoded fusion NA values to R NA - unfotunately FUSION uses -1.0 as NA values ..}

\item{dtm_file}{path to specific FUSION dtm file}

\item{ncore}{how many cores to use in processing dtm files - used with 'dir_dtm'}
}
\value{
raster, rasters, or dataframe with header(s) information
}
\description{
read a single USDA FUSION format raster into an r raster object (raster package).
read a single header
read headers for all files in directory
}
\details{
the DEM format is an esoteric raster format used only by (and required by) FUSION gridmetrics. SInce it is
required and we extensivey use FUSION (esp gridmetrics) we need the ability to interact with this file format.

\cr
Revision History
\tabular{ll}{
0.1 \tab 2014 July 09 - Operational \cr
}
}
\examples{

many_dtms=read_dtm_header(dir_dtm="c:\\\\temp\\\\dtm_files" ,out_file="c:\\\\temp\\\\headers.csv")

one_dtm=read_dtm_header("//dnrgpoly112/e$/rs_prototype/dtm/dtm_files/38_58.dtm")

}
\seealso{
\code{\link{read_las}}\cr
}
\author{
Jacob Strunk
}
