% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_las_cpp}
\alias{read_las_cpp}
\title{read_las_cpp}
\usage{
read_las_cpp(path, header_only = FALSE)
}
\arguments{
\item{path}{A file path}

\item{header_only}{If true, will only return the header portions.}
}
\value{
A two part list
}
\description{
A function to read LAS files.  Not really meant to be called by the user.
}
