% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{reader_las}
\alias{reader_las}
\alias{reader_las_coverage}
\alias{reader_las_circles}
\alias{reader_las_rectangles}
\title{Initialize the pipeline}
\usage{
reader_las(filter = "", ...)

reader_las_coverage(filter = "", ...)

reader_las_circles(xc, yc, r, filter = "", ...)

reader_las_rectangles(xmin, ymin, xmax, ymax, filter = "", ...)
}
\arguments{
\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. The available filters are those from LASlib and can be found by running \link{filter_usage}.
For a given algorithm when a filter is applied, only the points that meet the criteria are processes.
The most common strings are "-keep_first", "-keep_class 2", "drop_z_below 2". For more details see
\link{filters}.}

\item{...}{passed to other readers}

\item{xc, yc, r}{numeric. Circle centres and radius or radii.}

\item{xmin, ymin, xmax, ymax}{numeric. Coordinates of the rectangles}
}
\description{
This is the first stage that must be called in each pipeline. The stage does nothing and returns
nothing if it is not associated to another processing stage.
It only initializes the pipeline. \code{reader()} is the main function that dispatches into to other
functions. \verb{reader_las_*()} reads from LAS/LAZ files on disk.\code{reader_las_coverage()} processes the entire
point cloud. \code{reader_las_circles()} and \code{reader_las_rectangles()} read and process only some selected regions
of interest. If the chosen reader has no options i.e. using \code{reader_las()} it can be omitted.
}
\examples{
f <- system.file("extdata", "Topography.las", package = "lasR")

pipeline <- reader_las() + rasterize(10, "zmax")
ans <- exec(pipeline, on = f)

pipeline <- reader_las(filter = keep_z_above(1.3)) + rasterize(10, "zmean")
ans <- exec(pipeline, on = f)

# read_las() with no option can be omitted
ans <- exec(rasterize(10, "zmax"), on = f)
}
