% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{region_growing}
\alias{region_growing}
\title{Region growing}
\usage{
region_growing(
  raster,
  seeds,
  th_tree = 2,
  th_seed = 0.45,
  th_cr = 0.55,
  max_cr = 20,
  ofile = tempfile(fileext = ".tif")
)
}
\arguments{
\item{raster}{LASRalgoritm an algorithm producing a raster.}

\item{seeds}{LASRalgoritm an algorithm producing points used as seeds.}

\item{th_tree}{numeric. Threshold below which a pixel cannot be a tree. Default is 2.}

\item{th_seed}{numeric. Growing threshold 1. See reference in Dalponte et al. 2016. A pixel
is added to a region if its height is greater than the tree height multiplied by this value.
It should be between 0 and 1. Default is 0.45.}

\item{th_cr}{numeric. Growing threshold 2. See reference in Dalponte et al. 2016. A pixel
is added to a region if its height is greater than the current mean height of the region
multiplied by this value. It should be between 0 and 1. Default is 0.55.}

\item{max_cr}{numeric. Maximum value of the crown diameter of a detected tree \strong{(in data units)}.
Default is 20. \strong{BE CAREFUL} this algorithm exists in the \code{lidR} package and this parameter is in
pixels in \code{lidR}.}

\item{ofile}{character. Full outputs are always stored on disk. If \code{ofile = ""}
then the algorithm will not store the result on disk and will return nothing. It will however
hold partial output results temporarily in memory. This is useful for algorithm that are only
intermediate stage.}
}
\description{
Region growing for individual tree segmentation based on Dalponte and Coomes (2016) algorithm (see reference).
Note that this algorithm strictly performs segmentation, while the original method described in
the manuscript also performs pre- and post-processing tasks. Here, these tasks are expected to be
done by the user in separate functions.
}
\examples{
f <- system.file("extdata", "MixedConifer.las", package="lasR")

reader <- reader(f, filter = "-keep_first")
reader <- reader(f, filter = "-keep_first")
chm <- rasterize(1, "max")
lmx <- local_maximum(5)
tree <- region_growing(chm, lmx, max_cr = 10)
u <- processor(reader + chm + lmx + tree)

}
\references{
Dalponte, M. and Coomes, D. A. (2016), Tree-centric mapping of forest carbon density from
airborne laser scanning and hyperspectral data. Methods Ecol Evol, 7: 1236–1245. doi:10.1111/2041-210X.12575.
}
