% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_run_gridmetrics.R,
%   R/dev_run_gridmetrics2.R, R/dev_run_gridmetrics2b.R,
%   R/dev_run_gridmetrics3.R
\name{run_gridmetrics2}
\alias{run_gridmetrics2}
\alias{run_gridmetrics2}
\alias{run_gridmetrics2}
\alias{run_gridmetrics2}
\title{run gridmetrics across a project}
\usage{
run_gridmetrics2(lasR_project = NA, lasR_project_polys = NA,
  dir_out = "c:/temp/test_project/gridmetrics", n_core = 4,
  gridmetrics_type = c("c:\\\\fusion1\\\\gridmetrics.exe", "lasR"),
  heightbreak = 6, cellsize = 66, minht = 6, first = T, intensity = F,
  outlier = c(-5, 400), fusion_switches = "/nointensity /first",
  xmn = 561066, xmx = 2805066, ymn = 33066, ymx = 1551066,
  fun = compute_metrics2, temp = "c:\\\\temp\\\\run_gridmetrics\\\\",
  fast_cache = NA, n_cache = 90, dir_dtm = NA, dir_las = NA,
  skip_existing = T, con = NA, table = "gridmetrics",
  existing_coms = c(NA,
  "C:\\\\Temp\\\\run_gridmetrics\\\\2017Aug17_100740\\\\all_commands.txt"),
  debug = F, ...)

run_gridmetrics2(lasR_project = NA, lasR_project_polys = NA,
  dir_out = "c:/temp/test_project/gridmetrics", n_core = 4,
  gridmetrics_type = c("c:\\\\fusion1\\\\gridmetrics.exe", "lasR"),
  heightbreak = 6, cellsize = 66, minht = 6, first = T, intensity = F,
  outlier = c(-5, 400), fusion_switches = "/nointensity /first",
  xmn = 561066, xmx = 2805066, ymn = 33066, ymx = 1551066,
  fun = compute_metrics2, temp = "c:\\\\temp\\\\run_gridmetrics\\\\",
  fast_cache = NA, n_cache = 90, dir_dtm = NA, dir_las = NA,
  skip_existing = T, con = NA, table = "gridmetrics",
  existing_coms = c(NA,
  "C:\\\\Temp\\\\run_gridmetrics\\\\2017Aug17_100740\\\\all_commands.txt"),
  debug = F, ...)

run_gridmetrics2(lasR_project = NA, lasR_project_polys = NA,
  dir_out = "c:/temp/test_project/gridmetrics", n_core = 4,
  gridmetrics_type = c("c:\\\\fusion1\\\\gridmetrics.exe", "lasR"),
  heightbreak = 6, cellsize = 66, minht = 6, first = T, intensity = F,
  outlier = c(-5, 400), fusion_switches = "/nointensity /first",
  xmn = 561066, xmx = 2805066, ymn = 33066, ymx = 1551066,
  fun = compute_metrics2, temp = "c:\\\\temp\\\\run_gridmetrics\\\\",
  fast_cache = NA, n_cache = 90, dir_dtm = NA, dir_las = NA,
  skip_existing = T, con = NA, table = "gridmetrics",
  existing_coms = c(NA,
  "C:\\\\Temp\\\\run_gridmetrics\\\\2017Aug17_100740\\\\all_commands.txt"),
  debug = F, ...)

run_gridmetrics2(lasR_project = NA, lasR_project_polys = NA,
  dir_out = "c:/temp/test_project/gridmetrics", n_core = 4,
  gridmetrics_type = c("c:\\\\fusion1\\\\gridmetrics.exe", "lasR"),
  heightbreak = 6, cellsize = 66, minht = 6, first = T, intensity = F,
  outlier = c(-5, 400), fusion_switches = "/nointensity /first",
  xmn = 561066, xmx = 2805066, ymn = 33066, ymx = 1551066,
  fun = compute_metrics2, temp = "c:\\\\temp\\\\run_gridmetrics\\\\",
  fast_cache = NA, n_cache = 90, dir_dtm = NA, dir_las = NA,
  skip_existing = T, con = NA, table = "gridmetrics",
  existing_coms = c(NA,
  "C:\\\\Temp\\\\run_gridmetrics\\\\2017Aug17_100740\\\\all_commands.txt"),
  debug = F, ...)
}
\arguments{
\item{lasR_project}{csv file of intersections created by lasR_project() function}

\item{lasR_project_polys}{shape file of intersections created by lasR_project() function}

\item{existing_coms}{}

\item{lasR_project}{csv file of intersections created by lasR_project() function}

\item{lasR_project_polys}{shape file of intersections created by lasR_project() function}

\item{existing_coms}{}

\item{lasR_project}{csv file of intersections created by lasR_project() function}

\item{lasR_project_polys}{shape file of intersections created by lasR_project() function}

\item{existing_coms}{}

\item{lasR_project}{csv file of intersections created by lasR_project() function}

\item{lasR_project_polys}{shape file of intersections created by lasR_project() function}

\item{existing_coms}{}
}
\value{
<Delete and Replace>

<Delete and Replace>

<Delete and Replace>

<Delete and Replace>
}
\description{
<Delete and Replace>

<Delete and Replace>

<Delete and Replace>

<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2018-01-28 Header added \cr
}

<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2018-01-28 Header added \cr
}

<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2018-01-28 Header added \cr
}

<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2018-01-28 Header added \cr
}
}
\examples{

 gmi=run_gridmetrics2(
lasR_project_poly="D:\\\\projects\\\\2017_WA_DSM_Pilot_usgs\\\\2017Aug_NAIP_usgs\\\\lasR_project003.shp"
,dir_out="I:\\\\projects\\\\2017_WA_DSM_Pilot\\\\2017Aug_NAIP_usgs\\\\gridmetrics_07\\\\"
,dir_dtm="c:\\\\usgs_dtms\\\\dtms\\\\"
,dir_las="D:\\\\naip_2015_laz\\\\"
,n_core=10
,existing_coms="C:\\\\Temp\\\\run_gridmetrics\\\\2018Jan21_152618\\\\all_commands.txt"
,fast_cache=c(rep("r:\\\\temp",10),rep("c:\\\\temp",3),rep("i:\\\\temp",3),rep(NA,3))
,n_cache=400
)



 gmi=run_gridmetrics2(
lasR_project_poly="D:\\\\projects\\\\2017_WA_DSM_Pilot_usgs\\\\2017Aug_NAIP_usgs\\\\lasR_project003.shp"
,dir_out="I:\\\\projects\\\\2017_WA_DSM_Pilot\\\\2017Aug_NAIP_usgs\\\\gridmetrics_07\\\\"
,dir_dtm="c:\\\\usgs_dtms\\\\dtms\\\\"
,dir_las="D:\\\\naip_2015_laz\\\\"
,n_core=10
,existing_coms="C:\\\\Temp\\\\run_gridmetrics\\\\2018Jan21_152618\\\\all_commands.txt"
,fast_cache=c(rep("r:\\\\temp",10),rep("c:\\\\temp",3),rep("i:\\\\temp",3),rep(NA,3))
,n_cache=400
)



 gmi=run_gridmetrics2(
lasR_project_poly="D:\\\\projects\\\\2017_WA_DSM_Pilot_usgs\\\\2017Aug_NAIP_usgs\\\\lasR_project003.shp"
,dir_out="I:\\\\projects\\\\2017_WA_DSM_Pilot\\\\2017Aug_NAIP_usgs\\\\gridmetrics_07\\\\"
,dir_dtm="c:\\\\usgs_dtms\\\\dtms\\\\"
,dir_las="D:\\\\naip_2015_laz\\\\"
,n_core=10
,existing_coms="C:\\\\Temp\\\\run_gridmetrics\\\\2018Jan21_152618\\\\all_commands.txt"
,fast_cache=c(rep("r:\\\\temp",10),rep("c:\\\\temp",3),rep("i:\\\\temp",3),rep(NA,3))
,n_cache=400
)



 gmi=run_gridmetrics2(
lasR_project_poly="D:\\\\projects\\\\2017_WA_DSM_Pilot_usgs\\\\2017Aug_NAIP_usgs\\\\lasR_project003.shp"
,dir_out="I:\\\\projects\\\\2017_WA_DSM_Pilot\\\\2017Aug_NAIP_usgs\\\\gridmetrics_07\\\\"
,dir_dtm="c:\\\\usgs_dtms\\\\dtms\\\\"
,dir_las="D:\\\\naip_2015_laz\\\\"
,n_core=10
,existing_coms="C:\\\\Temp\\\\run_gridmetrics\\\\2018Jan21_152618\\\\all_commands.txt"
,fast_cache=c(rep("r:\\\\temp",10),rep("c:\\\\temp",3),rep("i:\\\\temp",3),rep(NA,3))
,n_cache=400
)


}
\seealso{
\code{\link{lasR_project}}\cr \code{\link{gridmetrics}}\cr

\code{\link{lasR_project}}\cr \code{\link{gridmetrics}}\cr

\code{\link{lasR_project}}\cr \code{\link{gridmetrics}}\cr

\code{\link{lasR_project}}\cr \code{\link{gridmetrics}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>

Jacob Strunk <Jstrunk@fs.fed.us>

Jacob Strunk <Jstrunk@fs.fed.us>

Jacob Strunk <Jstrunk@fs.fed.us>
}
