% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_las.R
\name{scan_las}
\alias{scan_las}
\title{scan las files and add them to summary table}
\usage{
scan_las(project = "some_project", project_year = "2099",
  proj4_name = "NAD_1983_HARN_StatePlane_Washington_South_FIPS_4601_Feet",
  proj4 = "1395 +proj=lcc +lat_1=47.33333333333334 +lat_2=45.83333333333334 +lat_0=45.33333333333334 +lon_0=-120.5 +x_0=500000.0001016001 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=us-ft +no_defs",
  dir_las = "", pattern = "[.]la.$", notes = "", create_polys = T)
}
\arguments{
\item{project}{name of lidar project}

\item{project_year}{year of lidar project}

\item{proj4_name}{human interpretable name of projection}

\item{proj4}{proj4 string for projection}

\item{dir_las}{where to find lidar files}

\item{pattern}{pattern to use in searching for las files}

\item{notes}{and descriptionn that may be helpful in using a project}

\item{create_polys}{output shapefiles of polygon bboxes}
}
\value{
<Delete and Replace>
}
\description{
scan las files and add them to summary table
}
\details{
scan las files and add them to summary table

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 March 08 Created\cr
}
}
\examples{

scan_las(project="test1", project_year="2015",dir_las="C:\\\\temp\\\\lidar_test\\\\",con=con_inv)

}
\seealso{
\code{\link{scan_dtm}}\cr \code{\link{read_las}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
