% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlas.R
\name{sectionParser}
\alias{sectionParser}
\title{Read LAS and split into raw sections
Intended to be used interally by the las function}
\usage{
sectionParser(las_dat)
}
\arguments{
\item{las_dat}{}
}
\value{
list. Raw sections of the las file.
}
\description{
Read LAS and split into raw sections
Intended to be used interally by the las function
}
