% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tiles.R
\name{select_tiles}
\alias{select_tiles}
\title{subset tiles in a lasR project}
\usage{
select_tiles(path_project, path_shapefile, ignore_proj4 = T, dir_out = "",
  return = F, plot = T)
}
\arguments{
\item{plot}{}
}
\description{
supply a polygon and subset tiles in a lasR project
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 April 14 Implemented \cr
}
}
\examples{
 tiles=select_tiles("C:\\\\projects\\\\2017_WA_DSM_Pilot\\\\DSM_Pilot_A_lasR\\\\intersections.csv"
 ,"C:\\\\projects\\\\2017_WA_DSM_Pilot\\\\boundary\\\\extentB.shp"
 ,dir_out="C:\\\\projects\\\\2017_WA_DSM_Pilot\\\\DSM_Pilot_A_lasR\\\\subset_project\\\\"
 )

}
\seealso{
\code{\link{another_function}}\cr \code{\link{yet_another_function}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
