% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{set_crs}
\alias{set_crs}
\title{Set the CRS of the pipeline}
\usage{
set_crs(x)
}
\arguments{
\item{x}{integer or string. EPSG code or WKT string understood by GDAL}
}
\description{
Assign a CRS in the pipeline. This stage \strong{does not} reproject the data. It assigns a CRS. This
stage affects subsequent stages of the pipeline and thus should appear close to \link{reader_las}
to assign the correct CRS to all stages.
}
\examples{
# expected usage
pipleine = reader_las() + set_crs(2949) + rasterize(10, "max")

# fancy usages are working as expected
pipleine = reader_las() + set_crs(1044) + rasterize(10, "max") + set_crs(1044) + rasterize(10, "max")
}
