% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{sort_points}
\alias{sort_points}
\title{Sort points in the point cloud}
\usage{
sort_points(spatial = TRUE)
}
\arguments{
\item{spatial}{boolean Add spatial sorting stage.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
The points are sorted by scanner channel, gpstime and return number in order to maximize the LAZ
compression. An optional second sort stage can be added to also sort points spatially. In this case
a grid of 50 meters is applied and points are sorted by scanner channel, gpstime and return number
within each cell of the grid. This increase data locality, seeds up spatial query but increases a
bit the final size of the files when compressed in LAZ format.
}
\examples{
#'
f <- system.file("extdata", "Topography.las", package="lasR")
exec(sort_points(), on = f)
}
