% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{splitHeaderLines}
\alias{splitHeaderLines}
\title{splitHeaderLines}
\usage{
splitHeaderLines(lines)
}
\arguments{
\item{lines}{Vector of LAS header strings}
}
\value{
A data.frame of parsed header values
}
\description{
Splits lines from an LAS header
}
