% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{splitSectionData}
\alias{splitSectionData}
\title{Splits a raw section into lines for further parsing}
\usage{
splitSectionData(dat)
}
\arguments{
\item{dat}{character. A raw LAS section}
}
\value{
character with section split into separate lines
}
\description{
Intended to be used internally by the las function
}
