% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempfile.R
\name{temporary_files}
\alias{temporary_files}
\alias{temptif}
\alias{tempgpkg}
\alias{tempshp}
\alias{templas}
\alias{templaz}
\alias{temppcd}
\title{Temporary files}
\usage{
temptif()

tempgpkg()

tempshp()

templas()

templaz()

temppcd()
}
\value{
string. Path to a temporary file.
}
\description{
Convenient functions to create temporary file with a given extension.
}
\examples{
tempshp()
templaz()
}
