% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trim_ws}
\alias{trim_ws}
\title{trim_ws}
\usage{
trim_ws(s)
}
\arguments{
\item{s}{A string}
}
\value{
A string with no whitespace
}
\description{
A function to trim the whitespace from the ends of a string
}
