% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readers.R
\name{versionParser}
\alias{versionParser}
\title{Read LAS file version information}
\usage{
versionParser(version_dat)
}
\arguments{
\item{version_dat}{character. Character vector containing the version section
of a LAS file.}
}
\description{
Read LAS file version information
}
