% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{versionWriter}
\alias{versionWriter}
\title{Creates a character vector of the LAS Version Information Section}
\usage{
versionWriter(las_version)
}
\arguments{
\item{las_version}{data.frame. Containing parsed LAS version information}
}
\value{
character.
}
\description{
Generates a character vector of the LAS Version Information Section, with
each new line as a separate element. This can be written to a file
using standard IO functions
}
