% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{wellParser}
\alias{wellParser}
\title{Parse well data}
\usage{
wellParser(well_dat)
}
\arguments{
\item{well_dat}{character. Raw section for well information}
}
\value{
list. List of well section information
}
\description{
Parse well data
}
