% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dtm.R
\name{write_dtm}
\alias{write_dtm}
\title{write USDA FUSION format raster}
\usage{
write_dtm(dtm, dir_out = NULL, NA_val = as.integer(-1), zfmt = NA)
}
\arguments{
\item{dtm}{raster file}

\item{NA_val}{recode fusion NA values to R NA values - unfotunately FUSION uses -1.0 as NA values ..}

\item{dtm_out}{path to specific FUSION dtm file}
}
\value{
Null, data is written to file
}
\description{
write USDA FUSION format raster
}
\details{
the DEM format is a niche raster format used only by (and required by) FUSION gridmetrics. SInce it is
required and we extensively use FUSION (esp gridmetrics) we need the ability to interact with this file format.

\cr
Revision History
\tabular{ll}{
0.1 \tab 2017 Dec 12 - Operational \cr
}
}
\examples{
 dtm1=read_dtm("C:\\\\Temp\\\\36_63.dtm")
 write_dtm(dtm=dtm1,dir_out="c:\\\\temp\\\\36_63_b.dtm")

}
\seealso{
\code{\link{read_las}}\cr
}
\author{
Jacob Strunk <jstrunk@fs.fed.us>
}
