% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{write_las}
\alias{write_las}
\title{Write LAS or LAZ files}
\usage{
write_las(
  ofile = paste0(tempdir(), "/*.las"),
  filter = "",
  keep_buffer = FALSE
)
}
\arguments{
\item{ofile}{character. Output file names. The string must contain a wildcard * so the wildcard can
be replaced by the name of the original tile and preserve the tiling pattern. If the wildcard
is omitted, everything will be written into a single file. This may be the desired behavior in some
circumstances, e.g., to merge some files.}

\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. For a given stage when a filter is applied, only the points that meet the criteria are processed.
The most common strings are \verb{Classification == 2"}, \code{"Z > 2"}, \code{"Intensity < 100"}. For more details see
\link{filters}.}

\item{keep_buffer}{bool. The buffer is removed to write file but it can be preserved.}
}
\description{
Write a LAS or LAZ file at any step of the pipeline (typically at the end). Unlike other stages,
the output won't be written into a single large file but in multiple tiled files corresponding
to the original collection of files.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
read <- reader()
tri  <- triangulate(filter = keep_ground())
normalize <- tri + transform_with(tri)
pipeline <- read + normalize + write_las(paste0(tempdir(), "/*_norm.las"))
exec(pipeline, on = f)
}
