% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{writeLAS}
\alias{writeLAS}
\title{Write an LAS object to a new LAS file}
\usage{
writeLAS(las, filename)
}
\arguments{
\item{las}{LAS object}

\item{filename}{character. File path}
}
\value{

}
\description{
Write an LAS object to a new LAS file
}
