% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMSAccess.R
\name{writeMSAccess_RODBC}
\alias{writeMSAccess_RODBC}
\title{Write data to MS access table - must be 32 bit windows}
\usage{
writeMSAccess_RODBC(db, nm, dat, dropTarget = F, doDebug = F)
}
\arguments{
\item{db}{access database}

\item{nm}{table name}

\item{dat}{data to write}

\item{dropTarget}{T/F drop target before writing ?}

\item{doDebug}{T/F enter browser() at beginning of function run}
}
\value{
<Delete and Replace>
}
\description{
handles some column name / type conversion issues
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 3/27/2020 Created \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{sqlSave}}\cr \code{\link{dbWritetable}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
