% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy2FIApoly.R
\name{xy2FIApoly}
\alias{xy2FIApoly}
\title{convert xy coordinates to fia plots}
\usage{
xy2FIApoly(idxy, names = c(plot = "plot", x = "x", y = "y"), create_polys = T)
}
\arguments{
\item{idxy}{dataframe with id, x, y columns available}

\item{names}{link dataframe columns to variable names used by funciton}

\item{create_polys}{TRUE / FALSE}
}
\value{
<Delete and Replace>
}
\description{
supply fia plot centers, and have have Spatial Polygons Dataframe returned, where each record is
 an FIA plot footprint: 4 fixed area subplots of 24 foot radius, a central subplot and 3 peripheral
 subplots 120 feet from plot center at 0,120,240 degrees.

 Ignores declination - assumes everything faces projected north
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab date and revisions.. \cr
}
}
\examples{

   res=xy2FIAplot(data.frame(plot=1:10, x=101:110*10000, y=101:110*10000),create_polys=T)
   spplot(res[1,],zcol=1,aspect=1,scales=list(draw=T),key.space=list(x=0.2,y=0.9,corner=c(0,1)))
   spplot(res)

}
\seealso{
\code{\link{bbox2polys}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
