% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{add_rgb}
\alias{add_rgb}
\title{Add RGB attributes to a LAS file}
\usage{
add_rgb()
}
\value{
If this stage transforms the point cloud in the pipeline it returns nothing. Otherwise
it returns the R object returned by the function 'fun'
}
\description{
Modifies the LAS format to convert into a format with RGB attributes. Values are zeroed: the underlying
point cloud is edited to be transformed in a format that supports RGB. RGB can be populated later
in another stage. If the point cloud already has RGB, nothing happens, RGB values are preserved.
}
\examples{
f <- system.file("extdata", "Example.las", package="lasR")

pipeline <- add_rgb() + write_las()
exec(pipeline, on = f)
}
