% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{classify_with_ipf}
\alias{classify_with_ipf}
\title{Classify noise points}
\usage{
classify_with_ipf(r = 1, n = 6L, class = 18L)
}
\arguments{
\item{r}{numeric. Radius of the sphere.}

\item{n}{integer. The maximal number of 'other points' in the sphere.}

\item{class}{integer. The class to assign to the points that match the condition.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
Classify points using Isolated Point Filter (IVF). The stage identifies points that have only a few other
points in their surrounding sphere neighborhood and edits the points to assign a target classification.
Used with class 18, it classifies points as noise. This stage modifies the point cloud in the pipeline
but does not produce any output.
}
