% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{classify_with_ivf}
\alias{classify_with_ivf}
\title{Classify noise points}
\usage{
classify_with_ivf(res = 5, n = 6L, class = 18L)
}
\arguments{
\item{res}{numeric. Resolution of the voxels. Can be a vector of 3 for x,y and z resolutions}

\item{n}{integer. The maximal number of 'other points' in the 27 voxels.}

\item{class}{integer. The class to assign to the points that match the condition.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
Classify points using Isolated Voxel Filter (IVF). The stage identifies points that have only a few other
points in their surrounding 3 x 3 x 3 = 27 voxels and edits the points to assign a target classification.
Used with class 18, it classifies points as noise. This stage modifies the point cloud in the pipeline
but does not produce any output.
}
