% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{delete_points}
\alias{delete_points}
\alias{delete_noise}
\alias{delete_ground}
\title{Filter and delete points}
\usage{
delete_points(filter = "")

delete_noise()

delete_ground()
}
\arguments{
\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. For a given stage when a filter is applied, only the points that meet the criteria are processed.
The most common strings are \verb{Classification == 2"}, \code{"Z > 2"}, \code{"Intensity < 100"}. For more details see
\link{filters}.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
Remove some points from the point cloud. This stage modifies the point cloud in the pipeline
but does not produce any output.
}
\examples{
f <- system.file("extdata", "Megaplot.las", package="lasR")
read <- reader_las()
filter <- delete_points(keep_z_above(4))

pipeline <- read + summarise() + filter + summarise()
exec(pipeline, on = f)
}
