% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{reader_las}
\alias{reader_las_coverage}
\alias{reader_las_circles}
\alias{reader_las_rectangles}
\title{Deprecated}
\usage{
reader_las(filter = "", select = "*", depth = NULL, ...)

reader_las_coverage(filter = "", select = "*", depth = NULL, ...)

reader_las_circles(xc, yc, r, filter = "", select = "*", depth = NULL, ...)

reader_las_rectangles(
  xmin,
  ymin,
  xmax,
  ymax,
  filter = "",
  select = "*",
  depth = NULL,
  ...
)
}
\arguments{
\item{filter}{see \link{reader}}

\item{select, depth}{see \link{reader}}

\item{...}{see \link{reader}}

\item{xc, yc, r}{see \link{reader}}

\item{xmin, xmax, ymin, ymax}{see \link{reader}}
}
\description{
Use \link{reader} instead.
}
