% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\name{exec}
\alias{exec}
\title{Process the pipeline}
\usage{
exec(pipeline, on, with = NULL, ...)
}
\arguments{
\item{pipeline}{a pipeline. A serie of stages called in order}

\item{on}{Can be the paths of the files to use, the path of the folder in which the files are stored,
the path to a \href{https://www.lutraconsulting.co.uk/blog/2023/06/08/virtual-point-clouds/}{virtual point cloud}
file or a \code{data.frame} containing the point cloud. It supports also a \code{LAScatalog} or a \code{LAS} objects
from \code{lidR}.}

\item{with}{list. A list of options to control how the pipeline is executed. This includes options to
control parallel processing, progress bar display, tile buffering and so on. See \link{set_exec_options}
for more details on the available options.}

\item{...}{The processing options can be explicitly named and passed outside the \code{with} argument.
See \link{set_exec_options}}
}
\description{
Process the pipeline. Every other functions of the package do nothing. This function must be
called on a pipeline in order to actually process the point-cloud. To process in parallel using
multiple cores, refer to the \link{multithreading} page.
}
\examples{
\dontrun{
f <- paste0(system.file(package="lasR"), "/extdata/bcts/")
f <- list.files(f, pattern = "(?i)\\\\.la(s|z)$", full.names = TRUE)

read <- reader_las()
tri <- triangulate(15)
dtm <- rasterize(5, tri)
lmf <- local_maximum(5)
met <- rasterize(2, "imean")
pipeline <- read + tri + dtm + lmf + met
ans <- exec(pipeline, on = f, with = list(progress = TRUE))
}
}
\seealso{
\link{multithreading} \link{set_exec_options}
}
