% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{normalize}
\alias{normalize}
\alias{hag}
\title{Height Above Ground (HAG)}
\usage{
normalize()

hag()
}
\description{
Normalize the point cloud using \link{triangulate} and \link{transform_with}. This process involves
triangulating the ground points and then using \code{transform_with} to linearly interpolate the elevation
for each point within the corresponding triangles. The \code{normalize()} function modifies the Z elevation
values, effectively flattening the topography and normalizing the point cloud based on Height Above Ground (HAG).
In contrast, the \code{hag()} function records the HAG in an extrabyte attribute named 'HAG', while preserving
the original Z coordinates (Height Above Sea Level).
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
pipeline <- reader() + normalize() + write_las()
exec(pipeline, on = f)
}
\seealso{
\link{triangulate}
\link{transform_with}
}
