% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{hulls}
\alias{hulls}
\title{Contour of a Delaunay triangulation}
\usage{
hulls(mesh = NULL, ofile = tempfile(fileext = ".gpkg"))
}
\arguments{
\item{mesh}{NULL or LASRalgorithm. A \code{triangulate} algorithm. If NULL take the bounding box of the
header of each file.}

\item{ofile}{character. Full outputs are always stored on disk. If \code{ofile = ""}
then the algorithm will not store the result on disk and will return nothing. It will however
hold partial output results temporarily in memory. This is useful for algorithm that are only
intermediate stage.}
}
\description{
This algorithm uses a Delaunay triangulation and computes its contour. The contour of a strict
Delaunay triangulation is the convex hull, but in lasR, the triangulation has a \code{max_edge} argument.
Thus, the contour might be a convex hull with holes.
}
\examples{
f <- system.file("extdata", "Topography.las", package = "lasR")
read <- reader(f)
tri <- triangulate(20, filter = "-keep_class 2")
contour <- hulls(tri)
pipeline <- read + tri + contour
ans <- processor(pipeline)
plot(ans)

}
\seealso{
\link{triangulate}
}
