% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/laser.metrics.R
\name{laser.metrics}
\alias{laser.metrics}
\title{Compute the number and proportion of returntypes}
\usage{
laser.metrics(x, id, gtv = 2, ctv = function(x) {     quantile(x[x >= gtv],
  probs = 0.95, type = 2) }, percentiles = seq(0.1, 0.9, 0.1), prefix = "",
  suffix = "", distpre = "H", denspre = "D")
}
\arguments{
\item{x}{surface height}

\item{id}{unique idenfitier for which variables will be aggregated}

\item{gtv}{ground threshold value. Default is 2}

\item{ctv}{canopy threshold value. Default is 95 percentile}

\item{percentiles}{the percentiles to be computed. Defaults to seq(0.1,0.9,0.1)}

\item{prefix}{optional prefix of varaible names}

\item{suffix}{optional suffix of varaible names}

\item{distpre}{indicator of distribution metrics. Default is "H"}

\item{denspre}{indicator of density metrics. Default is "D"}
}
\value{
data.frame with the numbers and proportion of echoes in differnt categories.
}
\description{
Read the header of a las-file into R
}
\note{
Date: Aug 26, 2011
}
\examples{
data(las)
prop <- laser.returns(las$dz,las$r,las$n,id = las$ID)
Echo.cat <- echo.cat(las$r,las$n)
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}

