% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laser.returntype.R
\name{laser.returntype}
\alias{laser.returntype}
\title{Classify returntypes, Echo categories}
\usage{
laser.returntype(RetNum, NumRet)
}
\arguments{
\item{RetNum}{return numb}

\item{NumRet}{number of returns}
}
\value{
vector with echo categories "SINGLE" (first and only), "FIRST" (first of many), "LAST"(last of many), "INTER" (intemediate), "ZERO" (labeling errors).
}
\description{
Classify echoes into echo categories based on return number and number of returns
}
\examples{
data(las)
Echo.cat <- laser.returntype(las$r,las$n)
table(Echo.cat)
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}
