% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{lasR-package}
\alias{lasR}
\alias{lasR-package}
\title{lasR: airborne LiDAR for forestry applications}
\description{
lasR provides a set of tools to process efficiently airborne LiDAR data in forestry contexts.
The package works with .las or .laz files. The toolbox includes algorithms for DSM, CHM, DTM, ABA,
normalisation, tree detection, tree segmentation, tree delineation, colourization, validation and
other tools, as well as a processing engine to process broad LiDAR coverage split into many files
efficiently.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lidar/lasR}
  \item Report bugs at \url{https://github.com/r-lidar/lasR/issues}
}

}
\author{
\strong{Maintainer}: Jean-Romain Roussel \email{jean-romain.roussel.1@ulaval.ca} [copyright holder]

Other contributors:
\itemize{
  \item Martin Isenburg (Is the author of the included LASlib and LASzip libraries) [copyright holder]
  \item Benoît St-Onge (Is the author of the included 'chm_prep' function) [copyright holder]
  \item Niels Lohmann (Is the author of the included json parser) [copyright holder]
  \item Volodymyr Bilonenko (Is the author of the included delaunator triangulation) [copyright holder]
  \item State Key Laboratory of Remote Sensing Science, Institute of Remote Sensing Science and Engineering, Beijing Normal University (Is the copyright holder of the included CSF) [copyright holder]
}

}
