% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasR-package.R
\docType{package}
\name{lasR-package}
\alias{lasR}
\alias{lasR-package}
\title{lasR: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
