% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasR.R
\name{lasR}
\alias{lasR}
\title{stuff}
\usage{
lasR()
}
\value{
stuff
}
\description{
stuff
}
\details{
stuff

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 APril 14 \cr
}
}
\examples{
stuff


}
\seealso{
\code{\link{another_function}}\cr \code{\link{yet_another_function}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
