% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasR.R
\name{lasR}
\alias{lasR}
\title{Read and process lidar}
\usage{
lasR()
}
\value{
#stuff
}
\description{
Read and process lidar
}
\details{
Read and process lidar

\cr
Revision History
\tabular{ll}{
1.0 \tab 2017 April 14 Created \cr
}
}
\examples{
#stuff

}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
\seealso{
\code{\link{lidR}}\cr
}

