% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmp.R
\name{ncores}
\alias{ncores}
\alias{half_cores}
\alias{sequential}
\alias{concurrent_files}
\alias{concurrent_points}
\title{Determine the number of available cores}
\usage{
ncores()

half_cores()

sequential()

concurrent_files()

concurrent_points()
}
\value{
An integer representing the number of available CPU cores.
}
\description{
\code{ncores()} returns the number of available CPU cores that can be utilized. \code{half_cores()}
returns half this number. \code{sequential()}, \code{concurrent_files()} and \code{concurrent_points()}
are convenient functions to assign a parallelization mode in \link{processor}.
}
