% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\name{processor}
\alias{processor}
\title{Process the pipeline}
\usage{
processor(pipeline, progress = FALSE, ...)
}
\arguments{
\item{pipeline}{a LASRpipeline. A serie of stages called in order (see examples)}

\item{progress}{boolean. Displays a progress bar.}

\item{...}{unused}
}
\description{
Process the pipeline. Every other functions of the package do nothing. This function must be
called on a pipeline in order to actually process the point-cloud. To process in parallel using
multiple cores, refer to the \link{multithreading} page.
}
\examples{
\dontrun{
f <- paste0(system.file(package="lasR"), "/extdata/bcts/")
f <- list.files(f, pattern = "(?i)\\\\.la(s|z)$", full.names = TRUE)

read <- reader(f, filter = "")
tri <- triangulate(15)
dtm <- rasterize(5, tri)
lmf <- local_maximum(5)
met <- rasterize(2, "imean")
pipeline <- read + tri + dtm + lmf + met

set_lasr_strategy(concurrent_files(4))
ans <- processor(pipeline, progress = T)
}
}
\seealso{
\link{multithreading}
}
