% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\name{processor}
\alias{processor}
\title{Process the pipeline}
\usage{
processor(pipeline, ncores = half_cores(), progress = FALSE, ...)
}
\arguments{
\item{pipeline}{a LASRpipeline. A serie of algorithms called in order}

\item{ncores}{integer. Number of cores to use. Some stages or some steps in some stages
are parallelised but overall one file is process at a time.}

\item{progress}{boolean. Displays a progress bar.}

\item{...}{unused}
}
\description{
Process the pipeline. Every other functions do nothing. This function must be called on a pipeline
to actually process the point-cloud
}
\examples{
\dontrun{
f <- paste0(system.file(package="lasR"), "/extdata/bcts/")
f <- list.files(f, pattern = "(?i)\\\\.la(s|z)$", full.names = TRUE)

read <- reader(f, filter = "")
tri <- triangulate(15)
dtm <- rasterize(5, tri)
lmf <- local_maximum(5)
met <- rasterize(2, mean(Intensity))
pipeline <- read + tri + dtm + lmf + met
ans <- processor(pipeline)
}
}
