% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\name{processor}
\alias{processor}
\title{Process the pipeline}
\usage{
processor(
  pipeline,
  ncores = half_cores(),
  mode = getOption("lasR.default_parallel_mode"),
  progress = FALSE,
  ...
)
}
\arguments{
\item{pipeline}{a LASRpipeline. A serie of stages called in order}

\item{ncores}{integer. Number of cores to use.}

\item{mode}{string. The parallelization mode. See details}

\item{progress}{boolean. Displays a progress bar.}

\item{...}{unused}
}
\description{
Process the pipeline. Every other functions do nothing. This function must be called on a pipeline
to actually process the point-cloud
}
\details{
There are 4 modes of parallel processing:
\describe{
\item{sequential}{No parallelization at all. \code{ncores = 1} not matter the actual user input.}
\item{concurent-points}{Point cloud files are process sequentially one by one. Inside the pipeline
some stages are parallelized and are able to process multiple points simultaneously. Not all stages
are natively parallelized.}
\item{concurent-files}{Files are process in parallel. Several files are loaded in memory
and processed simultaneously. The entire pipeline is parallelized but inside each stage
the points are process sequentially.}
\item{nested}{\strong{Not yet supported}. Files are process in parallel. Several files are loaded in memory
and processed simultaneously and inside some stages the points are process in parallel.}
}
\code{concurent-files} is likely the most desirable and fastest option on modern computers with
fast drive and many cores. However it uses more memory because it loads multiples files.
Also some stages do not support this type a parallelism because they call R code and R
is not multithreaded. For example a pipeline that implies the \link{callback()} stage
does not support \code{concurrent-files} multithreading because some R code is involved. The default
is \code{concurent-points} and can be changed globally using
\code{options("lasR.default_parallel_mode") = "concurent-files"}
}
\examples{
\dontrun{
f <- paste0(system.file(package="lasR"), "/extdata/bcts/")
f <- list.files(f, pattern = "(?i)\\\\.la(s|z)$", full.names = TRUE)

read <- reader(f, filter = "")
tri <- triangulate(15)
dtm <- rasterize(5, tri)
lmf <- local_maximum(5)
met <- rasterize(2, mean(Intensity))
pipeline <- read + tri + dtm + lmf + met
ans <- processor(pipeline)
}
}
\seealso{
\code{\link[=ncore]{ncore()}} \code{\link[=half_cores]{half_cores()}} \code{\link[=sequential]{sequential()}} \code{\link[=concurrent_points]{concurrent_points()}} \code{\link[=concurrent_files]{concurrent_files()}}
}
