% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.las.R
\name{read.las.helper}
\alias{read.las.helper}
\title{read.las.helper}
\usage{
read.las.helper(path, header_only = FALSE)
}
\arguments{
\item{path}{The path to the LAS file}

\item{header_only}{TRUE or FALSE depending on whether you want to return only header data.}
}
\value{
A two part list containing the LAS file and the header
}
\description{
Imports an LAS file to R
}
\keyword{internal}
