% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.las.R
\name{read.las.helper}
\alias{read.las.helper}
\title{read.las.helper}
\usage{
read.las.helper(path)
}
\arguments{
\item{path}{The path to the LAS file}
}
\value{
A two part list containing the LAS file and the header
}
\description{
Imports an LAS file to R
}
\keyword{internal}
