% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.las.R
\name{read.las}
\alias{read.las}
\title{read.las}
\usage{
read.las(path)
}
\arguments{
\item{path}{The path to the LAS file}
}
\value{
A two part list containing the LAS file and the header
}
\description{
Imports an LAS file to R
}
\examples{
#path = 'C:/Users/donal/Documents/R/3522229A.las'
#lasFile = read.las()
# http://pipeline.wyo.gov/whatups/whatuplas.cfm?nautonum=80871
}
