% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.las.R
\name{read.las}
\alias{read.las}
\title{read.las}
\usage{
read.las(paths, nthreads = 1, header_only = F)
}
\arguments{
\item{paths}{A vector of LAS file paths}

\item{nthreads}{An integer indicating the number of threads to use}

\item{header_only}{TRUE or FALSE depending on whether you want to return only header data.}
}
\value{
A list containing a Version, Well, and Curveset sections.
}
\description{
This function imports a LAS file to R.
}
\examples{
las = read.las(system.file("extdata", "Jonah_Federal_20-5.las", package = "lasr"))
}
