% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.las.R
\name{read.las}
\alias{read.las}
\title{read.las}
\usage{
read.las(paths, nthreads = 1)
}
\arguments{
\item{paths}{A vector of LAS file paths}

\item{nthreads}{An integer indicating the number of threads to use}
}
\value{
A two part list containing the LAS file and the header
}
\description{
Imports an LAS file to R
}
\examples{
las = read.las(system.file("extdata", "Jonah_Federal_20-5.las", package = "lasr"))
}
