% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.las.R
\name{read.las}
\alias{read.las}
\title{read.las}
\usage{
read.las(
  paths,
  nthreads = 1,
  header_only = FALSE,
  extra = FALSE,
  flatten = FALSE
)
}
\arguments{
\item{paths}{A vector of LAS file paths.}

\item{nthreads}{An integer indicating the number of threads to use.  If the number of 
threads and paths is greater than one it will run in parallel.  Default is one.}

\item{header_only}{TRUE or FALSE depending on whether you want to return only header data. 
If you only want headers, this will run faster.}

\item{extra}{Return the non-log data sections? FALSE by default.}

\item{flatten}{Return just the first set of log data from a file.  Makes it
easier to reference the log data since most files have only one set of log data.}
}
\value{
A list containing a Version, Well, and Curveset sections.
}
\description{
This function imports a LAS file to R.
}
\examples{
las = read.las(system.file("extdata", "Jonah_Federal_20-5.las", package = "lasr"))
}
