% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.las.R
\name{read.las}
\alias{read.las}
\title{read.las}
\usage{
read.las(paths, nthreads = 1, header_only = FALSE, logs_only = TRUE)
}
\arguments{
\item{paths}{A vector of LAS file paths.}

\item{nthreads}{An integer indicating the number of threads to use.  If the number of 
threads and paths is greater than one it will run in parallel.  Default is one.}

\item{header_only}{TRUE or FALSE depending on whether you want to return only header data. 
If you only want headers, this will run faster.}

\item{logs_only}{TRUE (default) if you only want to return up to the log section.}
}
\value{
A list containing a Version, Well, and Curveset sections.
}
\description{
This function imports a LAS file to R.
}
\examples{
las = read.las(system.file("extdata", "Jonah_Federal_20-5.las", package = "lasr"))
}
