% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.las.R
\name{read.las}
\alias{read.las}
\title{read.las}
\usage{
read.las(
  paths,
  nthreads = 1,
  header_only = FALSE,
  extra = FALSE,
  flatten = FALSE,
  pad_list = FALSE
)
}
\arguments{
\item{paths}{A vector of LAS file paths.}

\item{nthreads}{An integer indicating the number of threads to use.  If the number of 
threads and paths is greater than one it will run in parallel.  Default is one.}

\item{header_only}{TRUE or FALSE depending on whether you want to return only header data. 
If you only want headers, this will run faster.}

\item{extra}{LAS 3 files can have additional non-log data sections.  To import those
sections set this to TRUE.  By default it is FALSE because LAS 3 files with extra
data are rare in the wild.  Note that the extra data may have non-standard names so
referencing them may take some experimentation.}

\item{flatten}{If TRUE Return just the first set of log data from a file.
This makes it easier to reference the log data since most files have only 
one set of log data.}

\item{pad_list}{If TRUE adds a level to the resulting list when the number of 
files is one.  This makes automation easier since referencing the data will
always be consistent.  FALSE by default to make working with single files easier.}
}
\value{
A list containing a version, well, log, and path sections as well as 
any additional sections if extra is set to TRUE.
}
\description{
This function imports a LAS file to R.
}
\examples{
# load a single las file
f = system.file("extdata", "Jonah_Federal_20-5.las", package = "lasr")
las = read.las(f)

# get the version section
las$version

# get the well section
las$well

#' # get the original file path
las$path

# get the log data, definition, and parameters
las$log$log.1$data
las$log$log.1$definition
las$log$log.1$parameter

# use the flatten parameter to get just the first log data set
las = read.las(f, flatten = TRUE)

# referencing the log data is now shorter
las$log$data

# read multiple files.  consider setting nthreads > 1 if you have many files.
files = c(system.file("extdata", "Jonah_Federal_20-5.las", package='lasr'),
system.file("extdata", "las_3_cwls.las", package='lasr'))
las = read.las(files)

# in this case the imported las files are bound together in a list
# to reference the data you must first specify the element
# for instance, to get the log data from the second log
las[[2]]$log$log.1$data

# get extra data sections from a LAS 3 file
f = system.file("extdata", "las_3_cwls.las", package='lasr')
las = read.las(f, extra = TRUE)

# gets the core data as an example
las$core$core.1$data

# get just the header and not the actual data
# this is faster if you only need header data for many files
header = read.las(f, header_only=TRUE)

# using pad_list = TRUE so that accessing data from one file is the same
# as accessing it from many files (i.e. [[1]] is required)
las = read.las(f, pad_list = TRUE)
las[[1]]$log$log.1$data
}
