% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{reader}
\alias{reader}
\alias{reader_coverage}
\alias{reader_circles}
\alias{reader_rectangles}
\title{Initialize the pipeline}
\usage{
reader(files, filter = "", buffer = 0, ...)

reader_coverage(files, filter = "", buffer = 0, ...)

reader_circles(files, x, y, r, filter = "", buffer = 0, ...)

reader_rectangles(files, xmin, ymin, xmax, ymax, filter = "", buffer = 0, ...)
}
\arguments{
\item{files}{character. The path of the files to use. Supports a \code{LAScatalog} from \code{lidR}. Supports a path
to a folder.}

\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. The available filters are those from LASlib and can be found by running `print_filters()`.
For a given algorithm when a filter is applied, only the points that meet the criteria are processes.
The most common strings are "-keep_first", "-keep_class 2", "drop_z_below 2". For more details see
\link{filters}.}

\item{buffer}{numeric. Each file is read with a buffer. The default is 0, which does not mean that
the file won't be buffered. It means that the internal routine knows if a buffer is needed and will
pick the greatest value between the internal suggestion and this value if needed.}

\item{...}{Unused}

\item{x, y, r}{numeric. Circle centres and radius or radii.}

\item{xmin, ymin, xmax, ymax}{numeric. Coordinates of the rectangles}
}
\description{
This is the first algorithm that must be called in each pipeline. It specifies which files must be read.
The algorithm does nothing and returns nothing. It only initializes the pipeline. \code{reader()} is equivalent
to \code{read_coverage()}. \code{read_coverage()} reads all the files one by one, \code{read_circles()} and \code{read_rectangles()}
read only some selected regions of interest one by one.
}
\examples{
f <- system.file("extdata", "Topography.las", package = "lasR")
read <- reader(f)
ans <- processor(read)

}
