% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{sampling_voxel}
\alias{sampling_voxel}
\alias{sampling_pixel}
\alias{sampling_poisson}
\title{Sample the point cloud}
\usage{
sampling_voxel(res = 2, filter = "", ...)

sampling_pixel(res = 2, filter = "", ...)

sampling_poisson(distance = 2, filter = "", ...)
}
\arguments{
\item{res}{numeric. pixel/voxel resolution}

\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. For a given stage when a filter is applied, only the points that meet the criteria are processed.
The most common strings are \verb{Classification == 2"}, \code{"Z > 2"}, \code{"Intensity < 100"}. For more details see
\link{filters}.}

\item{...}{unused}

\item{distance}{numeric. Minimum distance between points for poisson sampling.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
Sample the point cloud, keeping one random point per pixel or per voxel or perform a poisson sampling.
This stages modify the point cloud in the pipeline but do not produce any output.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
read <- reader_las()
vox <- sampling_voxel(5)
write <- write_las()
pipeline <- read + vox + write
exec(pipeline, on = f)
}
