% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{sectionParser}
\alias{sectionParser}
\title{Read LAS and split into raw sections}
\usage{
sectionParser(las_dat)
}
\arguments{
\item{las_dat}{character. Raw LAS file as a character vector.}
}
\value{
list. Raw sections of the las file.
}
\description{
Parses a character vector representing a raw LAS file and splits it into
sections based on the header lines. Intended to be used internally by the las
function.
}
