% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{set_crs}
\alias{set_crs}
\title{Set the CRS of the pipeline}
\usage{
set_crs(x)
}
\arguments{
\item{x}{integer or string. EPSG code or WKT string understood by GDAL}
}
\description{
Assign a CRS in the pipeline. This stage \strong{does not} reproject the data. It assigns a CRS. This
stage affects subsequent stages of the pipeline and thus should appear close to \link{reader}
to assign the correct CRS to all stages.
}
\examples{
# expected usage
hmax = rasterize(10, "max")
pipeline = reader() + set_crs(2949) + hmax

# fancy usages are working as expected. The .tif file is written with a CRS, the .gpkg file with
# another CRS and the .las file with yet another CRS.
pipeline = set_crs(2044) + hmax + set_crs(2004) + local_maximum(5) + set_crs(2949) + write_las()
}
