% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{summarise}
\alias{summarise}
\title{Summary}
\usage{
summarise(zwbin = 2, iwbin = 50, metrics = NULL, filter = "")
}
\arguments{
\item{zwbin, iwbin}{numeric. Width of the bins for the histograms of Z and Intensity.}

\item{metrics}{Character vector. "min", "max" and "count" are accepted as well
as many others (see \link{metric_engine}). If \code{NULL} nothing is computed. If something is provided these
metrics are computed for each chunk loaded. A chunk might be a file but may also be a plot (see examples).}

\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. For a given stage when a filter is applied, only the points that meet the criteria are processed.
The most common strings are \verb{Classification == 2"}, \code{"Z > 2"}, \code{"Intensity < 100"}. For more details see
\link{filters}.}
}
\description{
Summarize the dataset by counting the number of points, first returns and other metrics for the \strong{entire point cloud}.
It also produces an histogram of Z and Intensity attributes for the \strong{entiere point cloud}.
It can also compute some metrics \strong{for each file or chunk} with the same metric engine than \link{rasterize}.
This stage does not modify the point cloud. It produces a summary as a \code{list}.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
read <- reader_las()
pipeline <- read + summarise()
ans <- exec(pipeline, on = f)
ans

# Compute metrics for each plot
read = reader_las_circles(c(273400, 273500), c(5274450, 5274550), 11.28)
metrics = summarise(metrics = c("z_mean", "z_p95", "i_median", "count"))
pipeline = read + metrics
ans = exec(pipeline, on = f)
ans$metrics
}
