% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{summarise}
\alias{summarise}
\title{Summary}
\usage{
summarise(zwbin = 2, iwbin = 25, filter = "")
}
\arguments{
\item{zwbin, iwbin}{numeric. Width of the bins for the histograms of Z and Intensity.}

\item{filter}{the 'filter' argument allows filtering of the point-cloud to work with points of
interest. The available filters are those from LASlib and can be found by running `print_filters()`.
For a given algorithm when a filter is applied, only the points that meet the criteria are processes.
The most common strings are "-keep_first", "-keep_class 2", "drop_z_below 2". For more details see
\link{filters}.}
}
\description{
Summarize the dataset by counting the number of points, first returns, classes. It also produces
a histogram of Z and Intensity. This algorithm does not modify the point cloud. It produces a
summary as a `list`.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")
read <- reader(f)
pipeline <- read + summarise()
ans <- processor(pipeline)
ans
}
