% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{tools}
\alias{tools}
\alias{print.LASRalgorithm}
\alias{print.LASRpipeline}
\alias{+.LASRpipeline}
\alias{c.LASRpipeline}
\title{Tools inherited from base R}
\usage{
\method{print}{LASRalgorithm}(x, ...)

\method{print}{LASRpipeline}(x, ...)

\method{+}{LASRpipeline}(e1, e2)

\method{c}{LASRpipeline}(...)
}
\arguments{
\item{x, e1, e2}{lasR objects}

\item{...}{lasR objects. Is equivalent to +}
}
\description{
Tools inherited from base R
}
\examples{
algo1 <- rasterize(1, "max")
algo2 <- rasterize(4, "min")
print(algo1)
pipeline <- algo1 + algo2
print(pipeline)
}
