% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{transform_with}
\alias{transform_with}
\title{Transform a point cloud using another stage}
\usage{
transform_with(stage, operator = "-", store_in_attribute = "")
}
\arguments{
\item{stage}{LASRpipeline. A stage that produces a triangulation or a raster.}

\item{operator}{string. '-' and '+' are supported.}

\item{store_in_attribute}{string. Use an extra bytes attribute to store the result.}
}
\value{
This stage transforms the point cloud in the pipeline. It consequently returns nothing.
}
\description{
This stage uses another stage that produced a Delaunay triangulation or a raster and performs an
operation to modify the point cloud. This can typically be used to build a normalization stage
This stage modifies the point cloud in the pipeline but does not produce any output.
}
\examples{
f <- system.file("extdata", "Topography.las", package="lasR")

# There is a normalize pipeline in lasR but let's create one almost equivalent
mesh  <- triangulate(filter = keep_ground())
trans <- transform_with(mesh)
pipeline <- mesh + trans + write_las()
ans <- exec(pipeline, on = f)

}
\seealso{
\link{triangulate}
\link{write_las}
}
