% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{write_vpc}
\alias{write_vpc}
\title{Write a Virtual Point Cloud}
\usage{
write_vpc(ofile)
}
\arguments{
\item{ofile}{character. The file path with extnsion .vpc where to write the virtual point cloud file}
}
\description{
Borrowing the concept of virtual rasters from GDAL, the VPC file format references other point
cloud files in virtual point cloud (VPC)
}
\references{
https://www.lutraconsulting.co.uk/blog/2023/06/08/virtual-point-clouds/
https://github.com/PDAL/wrench/blob/main/vpc-spec.md
}
