% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{write_vpc}
\alias{write_vpc}
\title{Write a Virtual Point Cloud}
\usage{
write_vpc(ofile, absolute_path = FALSE)
}
\arguments{
\item{ofile}{character. The file path with extension .vpc where to write the virtual point cloud file}

\item{absolute_path}{boolean. The absolute path to the files is stored in the tile index file.}
}
\description{
Borrowing the concept of virtual rasters from GDAL, the VPC file format references other point
cloud files in virtual point cloud (VPC)
}
\examples{
\dontrun{
pipeline = write_vpc("folder/dataset.vpc")
exec(pipeline, on = "folder")
}
}
\references{
\url{https://www.lutraconsulting.co.uk/blog/2023/06/08/virtual-point-clouds/}\cr
\url{https://github.com/PDAL/wrench/blob/main/vpc-spec.md}
}
